requirements.txt
python-dotenv>=1.0.1
flask>=3.0.2
oracledb>=2.1.2